# openapi.yaml
openapi: 3.1.0
info:
  title: Optica BFF API
  version: 0.1.0
  description: Generated minimal spec (partial) for core routes.
servers:
  - url: https://api.example.com
paths:
...
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Auth success }
        '401': { description: Invalid credentials }
  /auth/refresh:
    post:
      summary: Refresh access token
      responses:
        '200': { description: New tokens }
        '401': { description: Invalid or revoked }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key